mise à niveau

+++ src/UnTriangle.java
@@ -36,10 +36,10 @@ public class UnTriangle extends UnPolygone{
 		double y1 = this.getP1().getY() ;
 		double y2 = this.getP2().getY();
 		double y3 = this.getP3().getY() ;
-		double ABC = Math.abs (x1 * (y2 - y3) + x2 * (y3 - y1) + x3 * (y1 - y2));
-		double ABP = Math.abs (x1 * (y2 - p.getY()) + x2 * (p.getY() - y1) + p.getX() * (y1 - y2));
-		double APC = Math.abs (x1 * (p.getY() - y3) + p.getX() * (y3 - y1) + x3 * (y1 - p.getY()));
-		double PBC = Math.abs (p.getX() * (y2 - y3) + x2 * (y3 - p.getY()) + x3 * (p.getY() - y2));
+		int ABC = (int) (Math.abs (x1 * (y2 - y3) + x2 * (y3 - y1) + x3 * (y1 - y2)));
+		int ABP = (int) (Math.abs (x1 * (y2 - p.getY()) + x2 * (p.getY() - y1) + p.getX() * (y1 - y2)));
+		int APC = (int) (Math.abs (x1 * (p.getY() - y3) + p.getX() * (y3 - y1) + x3 * (y1 - p.getY())));
+		int PBC = (int) (Math.abs (p.getX() * (y2 - y3) + x2 * (y3 - p.getY()) + x3 * (p.getY() - y2)));
         return  (ABP + APC + PBC) == ABC;
     }
